SDK - Refactored command-line resolving
Moved the execution engine specific code to the component bridge.
